+++ Source/SwiftLintFramework/Extensions/Dictionary+SwiftLint.swift
@@ -45,6 +45,27 @@ extension Dictionary where Key: ExpressibleByStringLiteral {
         }
     }
 
+    var enclosedArguments: [[String: SourceKitRepresentable]] {
+        return substructure.flatMap { subDict -> [[String: SourceKitRepresentable]] in
+            guard let kindString = subDict["key.kind"] as? String else {
+                return []
+            }
+
+            switch SwiftVersion.current {
+            case .two:
+                guard SwiftDeclarationKind(rawValue: kindString) == .varParameter else {
+                    return []
+                }
+            case .three:
+                guard SwiftExpressionKind(rawValue: kindString) == .argument else {
+                    return []
+                }
+            }
+
+            return [subDict]
+        }
+    }
+
     var inheritedTypes: [String] {
         let array = self["key.inheritedtypes"] as? [SourceKitRepresentable] ?? []
         return array.flatMap { ($0 as? [String: String])?["key.name"] }